fix: fix cobalt and media command

+++ src/utils/config.ts
@@ -13,7 +13,7 @@ type CustomProxy = {
 export const Socks5Proxy: CustomProxy = {
     protocol: 'socks5',
     host: process.env.PROXY_HOST!,
-    port: parseFloat(process.env.PROXY_PORT!),
+    port: parseInt(process.env.PROXY_PORT!),
     auth: {
       username: process.env.PROXY_USER != undefined ? process.env.PROXY_USER : "",
       password: process.env.PROXY_PASS != undefined ? process.env.PROXY_PASS : ""
